refactor: move command registry to frontend/core (#7135)
move command registeration related logic out of infra module.
