+++ packages/common/infra/src/index.ts
@@ -1,7 +1,6 @@
 export * from './app-config-storage';
 export * from './atom';
 export * from './blocksuite';
-export * from './command';
 export * from './framework';
 export * from './initialization';
 export * from './livedata';
