+++ packages/frontend/core/src/commands/affine-help.tsx
@@ -1,10 +1,10 @@
 import type { useAFFiNEI18N } from '@affine/i18n/hooks';
 import { ContactWithUsIcon, NewIcon } from '@blocksuite/icons';
-import { registerAffineCommand } from '@toeverything/infra';
 import type { createStore } from 'jotai';
 
 import { openSettingModalAtom } from '../atoms';
 import { popupWindow } from '../utils';
+import { registerAffineCommand } from './registry';
 
 export function registerAffineHelpCommands({
   t,
