+++ packages/frontend/core/src/commands/affine-layout.tsx
@@ -1,9 +1,9 @@
 import type { useAFFiNEI18N } from '@affine/i18n/hooks';
 import { SidebarIcon } from '@blocksuite/icons';
-import { registerAffineCommand } from '@toeverything/infra';
 import type { createStore } from 'jotai';
 
 import { appSidebarOpenAtom } from '../components/app-sidebar';
+import { registerAffineCommand } from './registry';
 
 export function registerAffineLayoutCommands({
   t,
