+++ packages/frontend/core/src/commands/registry/README.md
