+++ packages/frontend/core/src/commands/registry/command.ts
