+++ packages/frontend/core/src/commands/registry/index.ts
