+++ packages/frontend/core/src/commands/registry/registry.ts
