+++ packages/frontend/core/src/modules/navigation/view/use-register-navigation-commands.ts
@@ -1,8 +1,8 @@
 import {
   PreconditionStrategy,
   registerAffineCommand,
-  useService,
-} from '@toeverything/infra';
+} from '@affine/core/commands';
+import { useService } from '@toeverything/infra';
 import { useEffect } from 'react';
 
 import { NavigatorService } from '../services/navigator';
