+++ packages/frontend/electron/package.json
@@ -68,7 +68,6 @@
     "react-router-dom": "^6.22.3",
     "rxjs": "^7.8.1",
     "semver": "^7.6.0",
-    "tinykeys": "patch:tinykeys@npm%3A2.1.0#~/.yarn/patches/tinykeys-npm-2.1.0-819feeaed0.patch",
     "tree-kill": "^1.2.2",
     "ts-node": "^10.9.2",
     "undici": "^6.12.0",
